+++ tests/core/server/test_dos.py
@@ -9,13 +9,14 @@ from chia.full_node.full_node_api import FullNodeAPI
 from chia.protocols import full_node_protocol
 from chia.protocols.protocol_message_types import ProtocolMessageTypes
 from chia.protocols.shared_protocol import Handshake
-from chia.server.outbound_message import make_msg, Message
+from chia.server.outbound_message import Message, make_msg
 from chia.server.rate_limits import RateLimiter
 from chia.server.server import ssl_context_for_client
 from chia.server.ws_connection import WSChiaConnection
 from chia.types.peer_info import PeerInfo
 from chia.util.errors import Err
 from chia.util.ints import uint16, uint64
+
 from tests.time_out_assert import time_out_assert
 
 log = logging.getLogger(__name__)
@@ -151,7 +152,7 @@ class TestDos:
         )
 
         # Construct an otherwise valid handshake message
-        handshake: Handshake = Handshake("test", "0.0.32", "1.0.0.0", 3456, 1, [(1, "1")])
+        handshake: Handshake = Handshake("test", "0.0.32", "1.0.0.0", uint16(3456), 1, [(1, "1")])
         outbound_handshake: Message = Message(2, None, bytes(handshake))  # 2 is an invalid ProtocolType
         await ws.send_bytes(bytes(outbound_handshake))
